requirements.txt
aiogram==2.21
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
certifi==2022.6.15
charset-normalizer==2.1.0
docopt==0.6.2
emoji-country-flag==1.3.0
frozenlist==1.3.0
h3==3.7.4
idna==3.3
multidict==6.0.2
numpy==1.23.1
pipreqs==0.4.11
pytz==2022.1
requests==2.28.1
timezonefinder==6.0.2
urllib3==1.26.10
yarg==0.1.9
yarl==1.7.2
loguru==0.6.0

python-dotenv~=0.20.0
